xdg-shell: Reword the xdg_wm_base introduction
Don't refer to things as "traditional desktop" as it is not defined
nor clear what that refers to; instead reword things in a more
non-explicit way. A reason for this is that xdg-shell is not strictly
meant only for traditional window stacking based desktop environments,
but should be equally suitable for stacking, tiling and potentially
other styles as well.

Signed-off-by: Jonas Ådahl <jadahl@gmail.com>
Reviewed-By: Mike Blumenkrantz <zmike@osg.samsung.com>